omniparser/pyproject.toml
[project]
name = "omniparser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.7",
    "comet-ml>=3.49.10",
    "datasets>=3.6.0",
    "faster-coco-eval>=1.6.6",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.3",
    "notebook>=7.4.2",
    "pillow>=11.2.1",
    "pycocotools>=2.0.10",
    "scipy>=1.15.3",
    "tensorboardx>=2.6.2.2",
    "torchmetrics>=1.7.1",
    "torchvision>=0.21.0",
    "transformers[torch]>=4.52.0",
    "ujson>=5.10.0",
]
